requirements.txt
django-generic-scaffold
django-markdown-app
django-user-accounts==2.0.0
pinax-webanalytics==2.0.4
pinax-eventlog==1.1.2

Django==1.10.3
pinax-theme-bootstrap==7.11.0
